+++ built_value_generator/lib/src/value_source_field.dart
@@ -11,6 +11,8 @@ import 'package:built_collection/built_collection.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value_generator/src/dart_types.dart';
 import 'package:built_value_generator/src/fields.dart' show collectFields;
+import 'package:built_value_generator/src/metadata.dart'
+    show metadataToStringValue;
 
 part 'value_source_field.g.dart';
 
@@ -43,7 +45,7 @@ abstract class ValueSourceField
 
   @memoized
   bool get isNullable => element.getter.metadata
-      .any((metadata) => metadata.constantValue?.toStringValue() == 'nullable');
+      .any((metadata) => metadataToStringValue(metadata) == 'nullable');
 
   @memoized
   bool get builderFieldExists => builderElement != null;